Refactor navbar
