+++ src/js/tools.js
@@ -3,8 +3,5 @@ export const createElement = ({tag, cls, txt}) => {
   element.classList.add(cls);
   element.textContent = txt;
 
-  // cls !== 'undefined' ? element.classList.add(cls) : null;
-  // text !== 'undefined' ? element.textContent = text : null;
-
   return element;
 }
